Fixed a bunch of bugs and added some stuff off of the trello list

+++ src/main/java/de/lucky44/luckybounties/files/config/CONFIG.java
@@ -100,7 +100,7 @@ public class CONFIG {
     public static String getString(String path){
         return config.getString(path);
     }
-    public static int getInt(String path) {return config.getInt(path);}
+    public static float getFloat(String path) {return config.getInt(path);}
 
     public static long toTickTime(String input){
         long out = 0;
